Replace cdr-table with cdr-data-table
