+++ package.json
@@ -23,6 +23,7 @@
     "@rei/cdr-card": "^0.2.0",
     "@rei/cdr-checkbox": "^1.0.0",
     "@rei/cdr-cta": "1.0.3-alpha.4",
+    "@rei/cdr-data-table": "0.1.0-alpha.1",
     "@rei/cdr-grid": "1.0.0",
     "@rei/cdr-icon": "^1.0.0",
     "@rei/cdr-img": "^1.0.1",
@@ -34,7 +35,6 @@
     "@rei/cdr-radio": "1.0.0",
     "@rei/cdr-rating": "1.0.0-alpha.3",
     "@rei/cdr-select": "0.1.1",
-    "@rei/cdr-table": "0.1.0-alpha.0",
     "@rei/cdr-tabs": "1.0.0-alpha.62",
     "@rei/cdr-text": "^1.0.1",
     "lodash": "^4.17.11",
